+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Fixed property naming validator issue checking unique constraint against account names, not property names.
 - Added article conversion/pageviews stats filtering based on selected property. GH-50
 - Added support for article tags (storing tags associated with articles, filtering in data tables). remp/remp#217
+- Added top articles endpoint for listing top articles by time, sections. remp/web#1010
 
 ### [Campaign]